implement patch from #424 improving ('fixing') our average latency calculation. Closes #424.

+++ pdns/syncres.hh
@@ -546,7 +546,7 @@ struct RecursorStats
   uint64_t nxDomains;
   uint64_t noErrors;
   uint64_t answers0_1, answers1_10, answers10_100, answers100_1000, answersSlow;
-  uint64_t avgLatencyUsec;
+  double avgLatencyUsec;
   uint64_t qcounter;
   uint64_t ipv6qcounter;
   uint64_t tcpqcounter;
